Update - add delete prompt & update styles

+++ script.js
@@ -6,7 +6,7 @@ import {
     replaceItem,
     changeItemStatus,
 } from './src/core.js';
-import { displayItems, addPrompt } from './src/ui.js';
+import { displayItems, addPrompt, confirmPrompt } from './src/ui.js';
 import data from './src/data.js';
 
 let tasks = [...data];
@@ -15,9 +15,13 @@ Object.values(listings).forEach((listing) => {
     listing.addEventListener('update', function (evt) {
         displayItems(listing, tasks);
     });
-    listing.addEventListener('click', function (evt) {
+    listing.addEventListener('click', async function (evt) {
         if (evt.target.matches('.js-delete')) {
             const id = Number.parseInt(evt.target.value);
+            const confirm = await confirmPrompt(); 
+            if (!confirm) {
+                return;
+            }
             tasks = deleteItem(id, tasks);
             listing.dispatchEvent(new CustomEvent('update'));
         }
